Minor bug fix: when writing the eventStream to the disk, the time-stamp was in shortened format, e.g., 6.234e+08

+++ src/openmsc.cc
@@ -334,7 +334,7 @@ void *sendStream(void *t)
 			mut.lock();
 			payload = (*eventMapIt).second;
 			if (streamToFileFlag)
-				file << eventMapIt->first << "\t" << payload << endl;
+				file << (int)eventMapIt->first << "\t" << payload << endl;
 			LOG4CXX_TRACE(logger, "Sending EventID " << payload << " at time " << (int)eventMapIt->first);
 			eventMap.erase(eventMapIt);
 			mut.unlock();
